+++ src/main/java/coma/spring/controller/QuestionController.java
@@ -58,78 +58,8 @@ public class QuestionController {
 		request.setAttribute("qdto", qdto);
 		return "question/questionView";
 	}
-	@RequestMapping("questionViewAdmin")
-	public String QuestionViewAdmin(HttpServletRequest request,int msg_seq)throws Exception{
-		QuestionDTO qdto = qservice.selectBySeq(msg_seq);
-
-		System.out.println("viewcount : "+qdto.getMsg_view());
-		request.setAttribute("qdto", qdto);
-		return "admin/admin_question_view";
-	}
 	
-	//관리자 페이지 1:1문의 리스트
-	@RequestMapping("AdminQuestion_list")
-	public String AdminQuestion_list(HttpServletRequest request)throws Exception{
-		MemberDTO mdto = (MemberDTO) session.getAttribute("loginInfo");
-		if(session.getAttribute("Aqcpage")==null) {
-			session.setAttribute("Aqcpage", 1);
-		}
-		try {
-			session.setAttribute("Aqcpage", Integer.parseInt(request.getParameter("Aqcpage")));
-		}catch(Exception e) {}
-		String Id=mdto.getId();
-		if(Id.contentEquals("administrator")) {
-			int Aqcpage = (int)session.getAttribute("Aqcpage");
-			List<QuestionDTO> qdto = qservice.selectByAdminQ(Aqcpage);
-			String navi = qservice.AdminQuestionNavi(Aqcpage);
-			
-			request.setAttribute("navi", navi);
-			request.setAttribute("list", qdto);
-			
-			return "admin/admin_question_list";
-		}else {
-			return "error";
-		}
-		
-	}
-	//1:1문의 답변페이지
-	@RequestMapping("questionAnswer")
-	public String QuestionAnswer(HttpServletRequest request,QuestionDTO qdto){
-		request.setAttribute("qdto", qdto);
-		System.out.println(qdto.getMsg_seq());
-		System.out.println(qdto.getMsg_sender());
-		return "admin/admin_qWrite";
-	}
-	//1:1문의 답변
-	@RequestMapping("questionAnswerSend")
-	public String questionAnswerSend(QuestionDTO qdto)throws Exception{
-		MemberDTO mdto=(MemberDTO)session.getAttribute("loginInfo");
-		
-		String admin = mdto.getId();
-		if(admin.contentEquals("administrator")) {
-			int answerSeq= qservice.getNextVal();
-			qdto.setMsg_seq(answerSeq);
-			int result = qservice.QuestionAnswer(qdto);
-			if(result==1) {
-				System.out.println(qdto.getMsg_view()+"번의 게시글에 대한 답변");
-				
-				QuestionDTO updto = new QuestionDTO();
-				updto.setMsg_seq(qdto.getMsg_view());
-				updto.setMsg_view(answerSeq);
-				System.out.println("답변업데이트 게시글 번호seq :"+qdto.getMsg_view()+"답변 게시글 번호 :"+answerSeq);
-				
-				int update = qservice.answerUpdate(updto);
-				
-				return "msg/msgWriteResult";
-			}else{
-				return "error";
-			}
-			
-			
-		}else {
-			return "error";
-		}
-	}
+
 	//1:1문의 보내기
 	@RequestMapping("insertQuestion")
 	public String insertQuestion(QuestionDTO qdto)throws Exception{
@@ -137,6 +67,7 @@ public class QuestionController {
 		String msg_sender = mdto.getId();
 		//보낸 사람 넣기
 		qdto.setMsg_sender(msg_sender);
+		qdto.setMsg_title("[1:1문의]"+qdto.getMsg_title());
 		int result = qservice.insertQuestion(qdto);
 		if(result==1) {
 			return "msg/msgWriteResult";
