+++ src/main/java/coma/spring/dto/QuestionDTO.java
@@ -1,6 +1,7 @@
 package coma.spring.dto;
 
 import java.sql.Timestamp;
+import java.text.SimpleDateFormat;
 
 public class QuestionDTO {
 	private int msg_seq;
@@ -12,6 +13,17 @@ public class QuestionDTO {
 	private int msg_view;
 	private int sender_del;
 	private int receiver_del;
+	private String date;
+	
+	public String getDate() {
+		SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd HH:mm");
+	    String date = sdf.format(msg_date);
+	    System.out.println(date);
+		return date;
+	}
+	public void setdate(String date) {
+		this.date=date;
+	}
 	
 	public QuestionDTO() {};
 	public QuestionDTO(int msg_seq,String msg_sender,String msg_receiver,String msg_title,String msg_text,Timestamp msg_date,int msg_view,int sender_del,int receiver_del) {
@@ -25,6 +37,7 @@ public class QuestionDTO {
 		this.msg_view=msg_view;
 		this.sender_del=sender_del;
 		this.receiver_del=receiver_del;
+		this.date=new SimpleDateFormat("yyyy-MM-dd-HH:mm").format(msg_date);
 	}
 	public int getMsg_seq() {
 		return msg_seq;
