+++ src/main/resources/mappers/question-mapper.xml
@@ -20,7 +20,7 @@
 		rnum from msg where
 		msg_sender = #{msg_receiver} and
 		msg_receiver = 'administrator' and
-		msg_title like '%[1:1문의]문의%') where rnum between
+		msg_title like '%[1:1문의]%') where rnum between
 		#{start} and #{end}
 	</select>
 	
@@ -33,7 +33,7 @@
 	<!-- 관리자 페이지 네비 카운트 -->
 	<select id="getAdminQCount" resultType="int">
 		select count (*) from msg
-		where msg_title like '%[1:1문의]문의%'
+		where msg_title like '%[1:1문의]%' and msg_sender != 'administrator'
 	</select>
 	
 	<!-- 답변에 view값 넣어주기 -->
@@ -46,14 +46,14 @@
 		select * from(select
 		msg.*,row_number() over(order by msg_date desc)
 		rnum from msg where
-		msg_title like '%[1:1문의]문의%') where rnum between
+		msg_title like '%[1:1문의]%' and msg_sender != 'administrator') where rnum between
 		#{start} and #{end}
 	</select>
 	<insert id="QuestionAnswer">
 		insert into msg values(#{msg_seq},'administrator',#{msg_receiver},'[1:1문의]답변입니다.',#{msg_text},sysdate,#{msg_view},0,0)
 	</insert>
 	<insert id="insertQuestion">
-		insert into msg values(msg_seq.nextval,#{msg_sender},'administrator','[1:1문의]문의글입니다',#{msg_text},sysdate,0,0,0)
+		insert into msg values(msg_seq.nextval,#{msg_sender},'administrator',#{msg_title},#{msg_text},sysdate,0,0,0)
 	</insert>
