+++ src/main/webapp/WEB-INF/views/question/writeQuestion.jsp
@@ -50,7 +50,7 @@ $(function(){
 			<tbody>
 				<tr align="center">
 					<th scope="col">제목</th>
-					<th scope="col">[1:1문의]문의글입니다.</th>
+					<th scope="col">[1:1문의]<input type="text" id="msg_title" name="msg_title"></th>
 				</tr>
 
 				<tr align="center">
