+++ src/main/webapp/resources/css/admin.css
@@ -32,7 +32,7 @@ body{
     width: 50%;
 }
 .menubtn{
-    height:10%;
+    height:8%;
     font-size: 2vw;
 
 }
